Refactor various handlers and services to improve code readability and maintainability. Changes include simplifying error handling, optimizing conditional checks, and enhancing JWT token management. Additionally, added integration tests for JWT access/refresh TTL checks in Docker environment.

+++ backend/src/services/email_verification_service.rs
@@ -461,7 +461,7 @@ impl EmailVerificationService {
             .jwt
             .mint_email_verification_token(user_id, tenant_id, ver_id)?;
         let expires_in = self.config.totp.email_verification_jwt_ttl_seconds;
-        Ok((ver_id, jwt, expires_in as u64))
+        Ok((ver_id, jwt, expires_in))
     }
 
     pub async fn resend_client_totp_enroll_email(
@@ -495,7 +495,7 @@ impl EmailVerificationService {
             .jwt
             .mint_email_verification_token(user_id, tenant_id, ver_id)?;
         let expires_in = self.config.totp.email_verification_jwt_ttl_seconds;
-        Ok((jwt, expires_in as u64))
+        Ok((jwt, expires_in))
     }
 
     /// Verify 6-digit code; does not set `email_verified` (user is already an account). Clears `cte:bind` Redis key.
